Update RadzenTooltip.razor (#289)
I believe the '@' shouldnt be there and if you cast the text to markupstring you dont need the renderfragment implementation? This would mean the overload from the tooltipservice should also be removed of course :-) I stumbled upon this because I implemented your tooltip in my base viewmodel and didnt want to start buildiing a fragment tree.