Dropdown reinitialization fix (Resolves #419) (#420)
* Update RadzenListBox.razor.cs

```item``` is a reference object, and this is not situable to be used as a key, because it is likely to change even though the properties of the object remain unchanged. Instead, we should use the ```ValueProperty``` as the key. This is most likely to be a primitive and also unique. If it isn't provided, the ```item``` is used (from what I understand about the ```PropertyAccess.GetItemOrValueFromProperty()``` implementation.

* Update RadzenDropDown.razor.cs

```item``` is a reference object, and this is not situable to be used as a key, because it is likely to change even though the properties of the object remain unchanged. Instead, we should use the ```ValueProperty``` as the key. This is most likely to be a primitive and also unique. If it isn't provided, the ```item``` is used (from what I understand about the ```PropertyAccess.GetItemOrValueFromProperty()``` implementation.