Update RadzenScheduler.razor.cs
Currently there's no easy way to get the current selectedView. This is because the selectedView that gets updated is the field, and not the public property.
The Scheduler.IsSelected is not an easy way to get it, as you need to test every view to get the current one.
Please consider updating the SelectedIndex property or making that SelectedView getter public.