Update RadzenDataGridColumn.cs
This method has only one reference, in the RadzenDataGrid.razor.cs. 
It has to change the ordering of the column based on the parameter.
So, checking the SortOrder value by GetSortOrder() isn't good, because it is the "past". The "order" value is the right SortOrder value.
See: https://forum.radzen.com/t/radzendatagrid-and-default-sortorder-and-datagridsettings-bug/13938